+++ src/main/java/sic/modelo/Pago.java
@@ -1,6 +1,7 @@
 package sic.modelo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -12,7 +13,7 @@ public class Pago implements Serializable {
     private Long id_Pago;    
     private long nroPago;
     private String nombreFormaDePago;    
-    private double monto;
+    private BigDecimal monto;
     private Date fecha;
     private String nota;
     private String nombreEmpresa;
